refactor: UserDataState를 preferences로 옮김
