+++ feature/onboarding/src/main/java/com/practice/onboarding/splash/SplashViewModel.kt
@@ -7,6 +7,7 @@ import com.google.firebase.auth.FirebaseAuth
 import com.practice.firebase.BlindarFirebase
 import com.practice.firebase.UserDataState
 import com.practice.preferences.PreferencesRepository
+import com.practice.preferences.UserDataState
 import com.practice.work.BlindarWorkManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.first
